fix: broke with a tie
In the old version of this program, the program would provide incorrect output when there was a tie (i.e. two sets of two days are both equally good days to buy and sell a stock). This old version of the program would output a relatively long array in those cases (e.g. stock_picker([3,1,31,3]) would output [1,3,0,2,4,1,3,0,2,4,1,3,0,2,4]. This updated version of the program uses hashes to fix that, so that only a single pair of days is given (one day to buy, another to sell).